refactor(dashboards): regroup the text panel's files
The kind had sixteen files flat in one folder. Split them the way the
sibling kinds are laid out: components in their own folders, and the
hooks and utils in the shared Panels folders that already hold their
peers. Renderer, definition and sections stay at the kind root.
