+++ frontend/src/pages/DashboardPage/DashboardContainer/Panels/kinds/TextPanel/components/CodeBlock/CodeBlock.module.scss
