+++ frontend/src/pages/DashboardPage/DashboardContainer/Panels/kinds/TextPanel/components/EditorPane/EditorPane.module.scss
