+++ frontend/src/pages/DashboardPage/DashboardContainer/Panels/kinds/TextPanel/components/ScrollToBottomPill/ScrollToBottomPill.tsx
