+++ frontend/src/pages/DashboardPage/DashboardContainer/Panels/utils/__tests__/interpolateVariables.test.ts
