+++ frontend/src/pages/DashboardPage/DashboardContainer/Panels/utils/interpolateVariables.ts
