+++ frontend/src/pages/DashboardPage/DashboardContainer/Panels/utils/syntaxLanguages.ts
